requirements.txt
louvain
numpy
python-igraph
tqdm
